+++ .gitignore
@@ -1,3 +1,1 @@
 node_modules
-.js/backup.js
-.js/scndbackup.js
