+++ index.html
@@ -15,7 +15,7 @@
         <section id="Books-List">
           
         </section>
-        
+
         <div class="horizontal-bar"></div>
         <main class="add-books">
             
@@ -28,6 +28,6 @@
               <button id="add-btn" type="button">Add</button>
             </div>
         </main>
-    <script src="js/index.js"></script>
+    <script src="js/index.js" type="module"></script>
     </body>
 </html>
