+++ js/Books.js
@@ -1,0 +1,7 @@
+// Create a class to save the Book Data
+export default class BookData {
+  constructor(author, title) {
+    this.author = author;
+    this.title = title;
+  }
+}
